+++ Engine/Source/Platform/Windows/Win32Window.h
@@ -15,8 +15,8 @@ namespace Pawn
 		virtual void SetEventCallback(const EventCallbackFunc& callback) override;
 
 	public:
-		virtual uint16 GetWidth() override;
-		virtual uint16 GetHeight() override;
+		virtual float32 GetWidth() override;
+		virtual float32 GetHeight() override;
 
 		inline HWND GetWindowHandle() { return m_Window; };
